+++ src/Service/ApiClient.php
@@ -5,7 +5,10 @@ namespace App\Service;
 use App\Exception\ParsingException;
 use App\Model\Application;
 use App\Model\Device;
-use App\Model\Gateway;
+use Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface;
+use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 final readonly class ApiClient
@@ -13,6 +16,7 @@ final readonly class ApiClient
     public function __construct(
         private HttpClientInterface $iotApiClient,
         private ApiParser $apiParser,
+        private int $gateWayOrgId,
     ) {
     }
 
@@ -28,10 +32,10 @@ final readonly class ApiClient
      * @throws ParsingException
      * @throws \DateInvalidTimeZoneException
      * @throws \DateMalformedStringException
-     * @throws \Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface
+     * @throws ClientExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws TransportExceptionInterface
      */
     public function getApplications(bool $filterOnStatus): array
     {
@@ -58,10 +62,10 @@ final readonly class ApiClient
      * @throws ParsingException
      * @throws \DateInvalidTimeZoneException
      * @throws \DateMalformedStringException
-     * @throws \Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface
+     * @throws ClientExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws TransportExceptionInterface
      */
     public function getApplication(int $id): Application
     {
@@ -83,10 +87,10 @@ final readonly class ApiClient
      * @throws ParsingException
      * @throws \DateInvalidTimeZoneException
      * @throws \DateMalformedStringException
-     * @throws \Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface
+     * @throws ClientExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws TransportExceptionInterface
      */
     public function getDevice(int $id): Device
     {
@@ -97,18 +101,27 @@ final readonly class ApiClient
     }
 
     /**
-     * @return array<Gateway>
+     * Retrieve a list of gateways.
      *
-     * @throws \Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface
-     * @throws \Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface
+     * @param bool $filterOnStatus
+     *   Indicates whether to filter gateways based on a specific status
+     *
+     * @return array
+     *   An array of parsed gateways
+     *
+     * @throws ParsingException
+     * @throws \DateInvalidTimeZoneException
+     * @throws \DateMalformedStringException
+     * @throws ClientExceptionInterface
+     * @throws RedirectionExceptionInterface
+     * @throws ServerExceptionInterface
+     * @throws TransportExceptionInterface
      */
     public function getGateways(bool $filterOnStatus): array
     {
         $response = $this->iotApiClient->request('GET', '/api/v1/chirpstack/gateway', [
             'query' => [
-                'organizationId' => 2,
+                'organizationId' => $this->gateWayOrgId,
                 'offset' => 0,
                 'limit' => 500,
             ],